Move workflow run related tests to better places
