+++ cli/step_activator_test.go
@@ -1,0 +1,125 @@
+package cli
+
+import (
+	"os"
+	"path/filepath"
+	"testing"
+
+	"github.com/bitrise-io/bitrise/models"
+	"github.com/bitrise-io/go-utils/pathutil"
+)
+
+func Test_activateStepLibStep(t *testing.T) {
+	tests := []struct {
+		name        string
+		stepIDData  models.StepIDData
+		wantVersion string
+		wantErr     bool
+	}{
+		{
+			name: "Major version lock",
+			stepIDData: models.StepIDData{
+				SteplibSource: "https://github.com/bitrise-io/bitrise-steplib.git",
+				IDorURI:       "xcode-archive",
+				Version:       "1",
+			},
+			wantVersion: "1.10.1",
+			wantErr:     false,
+		},
+		{
+			name: "Major version lock (long form)",
+			stepIDData: models.StepIDData{
+				SteplibSource: "https://github.com/bitrise-io/bitrise-steplib.git",
+				IDorURI:       "xcode-archive",
+				Version:       "1.x.x",
+			},
+			wantVersion: "1.10.1",
+			wantErr:     false,
+		},
+		{
+			name: "Minor version lock",
+			stepIDData: models.StepIDData{
+				SteplibSource: "https://github.com/bitrise-io/bitrise-steplib.git",
+				IDorURI:       "xcode-archive",
+				Version:       "2.3",
+			},
+			wantVersion: "2.3.7",
+			wantErr:     false,
+		},
+		{
+			name: "Minor version lock (long form)",
+			stepIDData: models.StepIDData{
+				SteplibSource: "https://github.com/bitrise-io/bitrise-steplib.git",
+				IDorURI:       "xcode-archive",
+				Version:       "2.3.x",
+			},
+			wantVersion: "2.3.7",
+			wantErr:     false,
+		},
+		{
+			name: "Patch version lock",
+			stepIDData: models.StepIDData{
+				SteplibSource: "https://github.com/bitrise-io/bitrise-steplib.git",
+				IDorURI:       "xcode-archive",
+				Version:       "2.3.2",
+			},
+			wantVersion: "2.3.2",
+			wantErr:     false,
+		},
+		{
+			name: "Invalid version lock",
+			stepIDData: models.StepIDData{
+				SteplibSource: "https://github.com/bitrise-io/bitrise-steplib.git",
+				IDorURI:       "xcode-archive",
+				Version:       "1.2.3.4",
+			},
+			wantVersion: "",
+			wantErr:     true,
+		},
+		{
+			name: "Latest version (not supported at the moment)",
+			stepIDData: models.StepIDData{
+				SteplibSource: "https://github.com/bitrise-io/bitrise-steplib.git",
+				IDorURI:       "xcode-archive",
+				Version:       "x.x.x",
+			},
+			wantVersion: "",
+			wantErr:     true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			tmpDir, err := pathutil.NormalizedOSTempDirPath("activateStepLibStep")
+			if err != nil {
+				t.Errorf("failed to create tmp dir: %s", err)
+			}
+
+			destination := tmpDir
+			stepYMLCopyPth := filepath.Join(tmpDir, "step-yml", "step.yml")
+
+			if err := os.MkdirAll(filepath.Dir(stepYMLCopyPth), 0777); err != nil {
+				t.Errorf("failed to create dir for step.yml: %s", err)
+			}
+
+			got, _, err := activateStepLibStep(tt.stepIDData, destination, stepYMLCopyPth, true)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("activateStepLibStep() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			if got.Version != tt.wantVersion {
+				t.Errorf("activateStepLibStep() got = %v, want %v", got.Version, tt.wantVersion)
+			}
+
+			if err != nil {
+				return
+			}
+			if exist, err := pathutil.IsPathExists(filepath.Join(destination, "step.yml")); err != nil || !exist {
+				t.Errorf("step not activate at: %s", destination)
+			}
+
+			if exist, err := pathutil.IsPathExists(stepYMLCopyPth); err != nil || !exist {
+				t.Errorf("step.yml not copied at: %s", stepYMLCopyPth)
+			}
+		})
+	}
+}
